Update Thingamabot.ino
"servo" is a reserved word if using the Servo library (which I think we want to do). So I changed the variable named "servo" to "servoID". Also, as I have never actually written software for a living, I am limiting my first ever Github pull request to this tiny change!